Fix LASK5 label truncation: preserve calibrated floats end-to-end
LASK5 data.values are calibrated finger-target floats in [0,1] (PROTOCOL.md
7.2), but the parser ran them through an int cast, collapsing values like
0.9942197 to 0 and flattening the label gradient. This corrupted training
labels and broke byte-compatibility with the PC CaptureWriter, which writes
the floats.

Make the label path Double throughout: LabelFrame.values, the LASK5 parser,
the CSV writer, and the recorder's manual-label vector. Manual UI labels (int
sliders) convert to Double at the boundary, so the Compose UI is unchanged.
Kotlin Double.toString matches Python str(float) for the [0,1] range, so the
golden CSV stays byte-identical (regenerated from the real PC writer, incl. a
0.9942197 case).

Verified live against lask5-01 (V4). All 50 unit tests pass.

+++ app/src/main/java/org/openmuscle/connect/capture/CaptureRecorder.kt
@@ -22,7 +22,7 @@ class CaptureRecorder(
 ) {
     /** When non-null, every sensor frame is labeled with this vector. */
     @Volatile
-    var manualLabel: List<Int>? = null
+    var manualLabel: List<Double>? = null
 
     // Volatile so the UI ticker can read live stats from a different thread than
     // the one writing rows.
